requirements-py-3-10.txt
absl-py==2.1.0
alt==0.0.1
anyascii==0.3.2
asttokens==2.4.1
astunparse==1.6.3
backports.tarfile==1.2.0
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
deep_sort @ git+https://github.com/team488/deep_sort@79af029bb17ce867f88c5f96564c3a9824afbdcb
defusedxml==0.7.1
depthai==2.30.0.0
devtools==0.12.2
distlib==0.3.8
docutils==0.21.2
ecdsa==0.19.0
exceptiongroup==1.2.2
executing==2.2.0
fast-histogram==0.14
filelock==3.16.1
filetype==1.2.0
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
gorilla==0.4.0
grpcio==1.70.0
grpcio-tools==1.70.0
h5py==3.13.0
huggingface-hub==0.29.3
humanfriendly==10.0
identify==2.6.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.4
jwcrypto==1.5.6
keras==3.5.0
keyring==25.4.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.5
langdetect==1.0.9
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nh3==0.2.18
nodeenv==1.9.1
ntcore==0.1.7
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.21.0
onnxslim==0.1.48
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
protobuf==5.29.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyarmor.cli.core==7.6.2
pyasn1==0.6.0
pycapnp==2.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pyflame==0.3.2
Pygments==2.18.0
pynetworktables==2021.0.0
pyparsing==3.1.2
pypdfium2==4.30.1
pyproject_hooks==1.2.0
pyrealsense2==2.55.1.6486
pytesseract==0.3.13
pytest==8.3.3
python-dateutil==2.9.0.post0
python-doctr==0.11.0
python-jose==3.3.0
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.2
readme_renderer==44.0
requests==2.32.3
requests-futures==1.0.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.2
rknn-toolkit2==2.3.0
robotpy-apriltag==2025.3.2.0
robotpy-wpimath==2025.3.2.0
robotpy-wpiutil==2025.3.2.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-fmm==2025.1.29
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.3
shapely==2.0.7
singleton==0.1.0
six==1.16.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.96
ultralytics-thop==2.0.8
urllib3==2.2.3
virtualenv==20.26.5
Werkzeug==3.0.4
wrapt==1.16.0
XTablesClient==5.7.9
zeroconf==0.145.1
zipp==3.20.2
